Added the Sphere Menu
I now have the option of having both the quad and sphere menus.

The quad menu is more traditional and simply stretches the panoramic
images onto a quad, but the sphere menu shows the 360 image that you
are going to go into when you pick it.

I think I will do a little effect later where the sphere you pick grows
and consumes you, so that you understand that you are going into the
image!