Added functionality to choose images with Gear VR
Grabbed a set of scripts from OSVR for interactivity and used them to
replicate the functionality that I’d built on the computer, to work in
VR.

This now means that when you run the app through the GearVR you can tap
on the side of the headset and pick one of the image sphere’s to change
the Skybox!

This has made it apparent that I need a reticle, and I’d still ideally
like to be able to switch into VR mode when I want to, rather than have
to start in it by default!