Added the Reticle so its now obvious where you are looking and the menu option that you are picking
I grabbed the same reticle implementation from OSVR which slots in with
the rest of how the code has been built.

The only problem is that I’m still getting a depth conflict between the
reticle and the spheres in the scene.

The next thing to do is get a sphere with its normals flipped (so that
I stop using the script to do it, which seems unreliable), then figure
out how to ensure there’s no depth conflicts!