Replaced Reverse-Normal script with sphere with actual normals flipped, and brought the Reticle's default position in front of all interactive elements to avoid depth conflict
Downloaded a sphere from TurboSquid with its UVs correctly set up and
flipped it inside-out so that the normals are back-to-front.

I had to add a Sphere Collider around the ImageSphere’s instead of a
Mesh Collider because the Mesh collider had its normals flipped along
with the Mesh and was part of the reason I was getting depth conflicts.

I then set the Reticle’s default position to be 4 metres in front of
the camera so that its always in front of everything, and it will just
move backwards when the user is looking at the direction of an
interactive item.

The main thing I want to understand better now before I feel like I
have a very solid basic prototype of the core-idea is how to switch in
and out of VR mode - at the moment the app appears to crash or just go
black, when I try to switch out of VR mode.