Added AWS S3 support for back-end!
Imported the AWSSDK.S3 unitypackage and got an example of it working
where I download an object off the S3 server and map it onto one of my
ImageSphere prefabs!

The next thing to do on this front is to get that ImageSphere to be
selectable, and to switch the actual skybox with the new texture.

After that I’ll be able to make a minimal app that essentially works
the same as the current one with my 6 stock images, but they are all
downloaded at runtime rather than embedded in the app!