Made Test_02 such that it downloads the images from S3 then sets them on the ImageSpheres!
I now have a working example of a test where my 6 stock images are
actually downloaded from the server. Hence its a minimal app, and the
images are grabbed at runtime, then they replace the SkyboxSphere and
give the same illusion of the user being within the 360 image as before!

The key things that will need to improve in this workflow are:
1) The ImageSphere is used as a Skybox instead of the skybox in
RenderSettings, because I don’t know how to create the cube map on the
fly - this leads to the visual quality of the images noticeably
dropping.
2) There is a stutter at the moment that the Image is Loaded onto the
Sphere that interferes with the main thread, this operation needs to be
done Asynchronously to not interrupt the user experience
3) Whilst imageSpheres are being updated they need to show that some
sort of loading is going on, because atm its not obvious at all