Gallery button correctly opens and loads first 6 photos + Menu Swipe is now correctly being registered based on length of swipe
The code to open the gallery and set the ImageSpheres according to the
pictures found in the android devices gallery now works for the first 6
images! I’ve also written code for transitioning to the next set of 6
pictures but I haven’t yet tested it or added UI for it.

Essentially I’m fishing out all the file paths in the
"/storage/emulated/0/DCIM/Gear 360/" directory, adding them to a list,
then loading the data into a texture and setting that on the
ImageSphere, up until I run out of ImageSpheres (there’s only 6 in the
scene).

The main problems with this method are (1) that I’ve hardcoded the file
path, and (2) that I’m doing it on the main thread so the whole
application freezes for a sec when this happens (I think/hope running
the code in a co-routine will probably solve this problem).

The problem with Swiping was that the code would previously normalise
the difference in X and Y position (which probably makes sense for
other applications) meaning that it would always successfully register
as an up/down swipe regardless of the SwipeWidth value I put in the
editor.