Attempted at removing the Synchronous block for Image Loading - alleviated it slightly, but didn't fix the problem
I believed Coroutines may be the solution for removing the synchronous
block on image loading as I thought that they were Unity’s solution to
threading.

However, coroutines and yielding is only a way of giving control back
to the main thread to continue with whatever other work it needs to do,
for it to then come back later and continue where it was running from,
therefore they do not take away the work from the main thread.

I therefore moved into looking at System.Threading.Thread, which turns
out to be the correct way to do Threading in a .NET environment, but
found out a sad truth about Unity - it doesn’t really like
multithreading!

So although I was able to get a separate thread running midway through
execution, I wasn’t able to offload the bulky work to this thread!
Neither Texture.LoadImage() nor new WWW(“file://“ + path) were able to
work on a separate thread, they needed to be run on the main thread =(

(As a side point, this is where I found out that “new WWW(“file://“ +
path)“ does indeed work, and it probably wasn’t working before because
of the “Write Access: External (SDCard)” setting, as I had struggled
with when I tried LoadImage() )

So I went back to coroutines and did a temporary hack. Between every
incredibly slow and blocking load (sad face), I stick a yield for 3
seconds, so that at least it doesn’t feel like its stuck for ages, it
simply judgers as the images are downloaded.

In the long-run I NEED to figure out how to do some equivalent of
Asynchronous Image loading, otherwise this concept will simply not work.

The next thing I’m going to try and look at though is the fact that
after I’ve scrolled a couple of times to the right through the gallery,
the app will crash, and that’s another pretty big problem!