Added functionality for scrolling along downloaded images from the S3 server
The user can now scroll through all the pictures that he’s uploaded.
This completes the pipeline of taking a picture in the real world,
opening it in the viewing gallery, uploading it to the cloud, viewing
it in your profile!

I now need to wrap this up into a single app such that its clear to the
user when he’s viewing the Gallery and when he’s in his Profile,
because it wasn’t clear to Simone and Bruno, and if they don’t get it
then I’m in trouble. This will involve improving the UI a little but I
don’t want to get bogged down in doing UI just yet.

I do also need to make the fact that buttons are being pressed more
obvious because users are currently unaware when a button gets pushed
in or not.

The biggest notable problems still visible for me are that the
resolution is still not great (I need to look into the cube mapping at
some point), and more importantly that image loading freezes the entire
app, as its not asynchronous.