MenuButton can now flip between images when they are pressed down
There was a real lack of feedback before when a button was pressed, so
I’ve added functionality into the MenuButton script that can now handle
image switching when a button switches between being help down to up.

I’ve also created corresponding Up/Down images for all the button icons
that I’ve got in my scenes. Therefore its now the case that you get
very clear feedback from pushing on buttons in a scene.

Next I need to make it such that when the texture on an ImageSphere is
swapped, its also very clear. I think I probably need to do some sort
of visual animation of a sphere disappearing and a new one coming in.