Created a new 4th Scene that merges both the Profile and Gallery
Firstly I changed the UI setup such that the two central spheres now
sit at the bottom and the Home (Profile) and Gallery buttons now appear
at the top with the VReel Logo.

In doing this I also moved around the hierarchy splitting the
ImageSpheres away from all the MenuButtons, then making the MenuBar,
the Gallery options and the Profile options all separate too!

By doing this I was able to then give the MenuController a little more
responsibility and have it switch the relevant Options on/off depending
on the MenuBar option selected. I also created a SelectedButton script
that goes on these special MenuBar buttons (atm its just the Gallery
and Profile buttons), which the MenuController knows about and calls
through to actually perform this functionality.

I also quickly changed it such that the images coming up are reversed,
so that we get newest images first, and older images as we scroll with
the arrows.

The main problems at the moment are the loading freezes (which are
still incredibly painful), and some weird bug that’s cropped up with
the depth of the Reticle, which seems to be conflicting with the depth
of the other icons!