Disabled VRInput firing certain events when you are not in VR!
I check UnityEngine.VR.VRSettings.enabled so that we don’t start
calling through to MenuButtons when VR is disabled without pressing
them intentionally.

The main problem I was trying to fix was that any button that was in
the centre of the camera would automatically be pressed because that’s
where the reticle sits by default!