Fixed the issue with the reticle drawing under the Menu buttons
The problem was an ordering issue. Apparently Sprites/Images don’t draw
to the z-buffer at all. So there’s no guarantee of what will be drawn
first.

However, I changed the GUIReticle to use the SpriteRenderer component
instead of an Image script, because the SpriteRenderer has Layers to
it, and I changed the layer that it was using, which fixed the problem.

This feels like a bit of a hack as the rest of the UI elements are
still using the Image script. I’m thinking that in the long-run I need
to either pick using SpriteRenderer all the time or just Image scripts
along with the CanvasRenderer. I honestly don’t see why I need the
Canvas though and if I don’t then I’ll stick to just having
SpriteRenderers do all the work!