Added a CoroutineQueue in reducing the loading lag + Made Arrows and ImageSpheres disappear/appear appropriately
I found the coroutine queue on stack overflow and its great because it
means that I’m not having several jobs running at once, and it allows
me to have a better handle on what I’m loading at any moment in time!
So when a user presses the right/left arrow buttons then the queue can
throw away all the jobs it was meant to do, focus on the new jobs it
has been given.

The arrows are always updating now to check if the user has reached the
end of the gallery or profile, and if they have they will disappear
appropriately. Similarly when reaching the end of the gallery/profile,
any spare ImageSpheres that are not being used, will be hidden giving
the impression that you’ve reached the end.

I’ve also made the executive decision to have only 5 ImageSphere’s at a
time. I actually did this partly because I was running out of memory
(although the memory issue still happens with 3 spheres and I think is
to do with the actual LoadImage() call), but also because I think it
looks better.

Although the coroutinequeue alleviates the problem, the main problem I
still have is that Loading the Image freezes the app, as a large
frameout occurs. I’ve posted a couple of questions on StackOverflow but
I will continue to look at removing this freeze because its the biggest
problem at the moment!