Made the MenuController disable Meshes, Images and Colliders instead of a full entity subtree
As I was disabling the full Menu subtree, I was causing a lot of
scripts to stop running, which was not ideal. I’ve now made it such
that the visual elements of the menu and the colliders switch off when
you swipe appropriately, but the scripts will continue running.

I’ve also fixed a bug with the swiping where the X input was being used
for Vertical and Y input for Horizontal recognition. This has now been
correctly flipped and fixed, and I’ve upped the required swipe to be
100 pixels in length.

The overall user experience is a lot better in general now, the key
problem is still the freezing from the texture copy construction, and
the crashing every so often, which is also tied to copying/loading
textures!