Added/Created a working VR Keyboard Prefab!
I’m trying to write as little code as possible and build things in an
agile way so I actually grabbed a keyboard from the Asset Store that I
payed $12 for.

Once I had it, I realised that it was built in a pretty simple way, but
the Keyboard didn’t work immediately with my setup. So I built some
prefabs and made it such that it did work with my VRInteractiveItem and
MenuButtons.

This required tidying MenuButton a little - which has now become a lot
more powerful and generic, since it has a UnityEvent and can call
through into functions that I hook up through the Editor, plus it can
deal with Mousing Over and switching images accordingly. With this I
actually deleted some scripts that I had before, which I don’t need as
I can just set the function to be called through the UI with the help
of UnityEvents.

There was also a good chunk of monotonous, essentially copy and paste
work, of prefab-ing up each Button on the keyboard and updating it to
work accordingly.

I’ve ended up with a bunch of extra assets in my project that I don’t
need although I’ll delete them with time. Most importantly though, I’ve
now got a working VR Keyboard that I can use anytime I want in the App!
And at some point if I decide to update it, I now have a base to work
off! I’m pretty pleased about this as plenty parts of the App will
require a VR Keyboard not just the Login screen!