Fixed Keyboard in VR as the 2D Colliders weren't working
I tidied the Login scene a little bit then tested the new build and
found that the Keyboard wasn’t interacting at all.

Turned out that the problem was that I was using 2D Box Colliders
(which I assumed would be fine), but they didn’t work with the VR
Interaction system, so I changed it to be a standard (3D) Box Collider
and that fixed the keyboard!