Improved Keyboard Script and Functionality + Tidied project a bit
The Keyboard Script will now hide default text (eg. Username +
Password), display asterixes for passwords, and is cleaner in terms of
naming convention (eg. using “m_” for member variables).

I also tidied the project by moving certain assets into folders, and
changed the Profile Icon to be more inline with the Profile icon on
Instagram, instead of a Home icon as it was before!