Created the Basic Login Flow and accompanying script
A new script called “LoginFlow” controls the flow when clicking on the
buttons in the 3 Login Pages and interfaces with the AWSS3Client script
so that it can check whether: (1) Login is valid; (2) An email address
is not yet taken; (3) Sign up a new user!

The basic flow has 3 screens, the first giving Login and Sign Up
options, the next giving email availability checking or back, and the
final sign up confirmation or back. I’ve therefore also created new
Confirm and Back images for the appropriate buttons!