Fixed a bug where clicking the buttons in the Menu Bar would infinite loop
NOTE: This is the first commit since moving to the private repository!

The SelectedButton would call MenuButton’s SetSpriteButton(), which
would call through to HandleUp() and through to OnButtonSelected()
looping back to the SelectedButton doing its thing. Now I’ve created a
RefreshButtonSprite() function so that we aren’t calling HandleUp()
unnecessarily and breaking the infinite loop!