Added Error messages to LoginFlow and Imported some new AWS Unity Packages
I now have some of the basic error messages that I’ll need for the
Login Flow, these are somewhat hooked and functional except for the
fact that the actual code communicating with AWS’s Cognito to check
whether a user exists or not hasn’t been written yet.

When looking at how I was meant to communicate with AWS’s Cognito I
imported a bunch of probably useless unity packages (that I will delete
later in a full tidy up if indeed unnecessary).

More importantly though, I realised that there doesn’t seem to be a
unitypackage for me to import to use Cognito Identity, and I also can’t
find any sample code for using it in the AWS documentation - this has
me a little concerned that there’s no way to actually achieve user
log-in through Unity yet… this would be pretty bad if it were the case!

I’ve sent an email to AWS’s technical support to check if this is true,
I’m not sure how I’ll get around it if there’s no way to make users
using AWS with Unity…