Added a Welcome message when user logs in!
This is a simple thing that could easily be dismissed as irrelevant,
but I think its an important touch.

It shows me that I’m logging into the correct account, and makes the
app feel more friendly to the user. Plus, I’ve learned how to grab
things off Facebook if necessary.

Currently I change the text on a button, which says “Welcome
*username*!” and you can click on it for it to go away. Later I’ll make
this just sort of animate in and fade away by itself.

I also added a logout button, although I just disabled it and didn’t
write any functionality to make it work, as its not really my priority
at the moment, and I don’t intend on telling people to logout from my
app.

Next I need to work on the policy for the IAM role that a user takes on
when they get authenticated!