Updated the Login/Download/Upload to use Cognito!
The AWS support technician Chris came back to me with the solution for
uploading to S3 using Cognito correctly and I simply implemented it.

Nothing in principal will have changed to how the app works (although
I’ve noticed a bug where sometimes the images aren’t loading straight
away when you log into the app), but in the background we now
authenticate and store images using the Cognito ID (as opposed to the
Facebook one), which is great because this means if at some point I
move on to authenticating through other methods (eg. Twitter), then it
will be seamless to migrate to.

I also now prefix the Date and Time to any file I upload, so that I can
correctly display it in order, because S3 doesn’t have any way of
querying data based on time it was created. It also means that I can
upload the same image twice (as before uploading the same image would
just overwrite the previous one because of name clashes.