Bug Fixing!
1) The Login page now waits until the S3 client has been initialised
before it opens up, this removes the case where we’d leave the Login
page and the S3 client would not yet be valid to query, so we wouldn’t
get any images showing up.

2) Hiding the menu now correctly hides the scrolling arrows. This bug
cropped up because I moved the MenuController script to a different
entity, and the script queries the MenuController to know when to hide
itself.

3) Created an ImageSphereController class as an interface to
ImageSpheres, which now hide themselves when switching between Gallery
and Profile, to make it clearer that we are switching between modes!