Added SplashScreen and Icon + Fixed Selected Sprite issue + Made App Landscape only
I’ve made a very simple splash screen in 2D (which I doubt works in VR,
i’ll need to check), and I’ve used the VReel icon as the App Icon.

I’ve fixed a silly issue with MenuBarButtons remaining selected even
when they are not, and ensured that when a button has been pressed
once, it cannot be pressed multiple times (as this was causing all
sorts of weird issues) - essentially I just block the
RequestGalleryState/RequestProfileState in the AppDirector if you are
already in that state.

I also changed the orientation to be always Landscape.Right if you are
not in VR, because it was annoying when debugging to have it flicker
between orientations!