Addied Loading to Loggin in and Uploading images + Removed a bug with the background image switching by itself
1) When the user clicks the sign in button on Facebook it now brings up
a losing icon which also blocks access to all UI behind it (I simply
put a big old box collider in front of everything). This also appears
when uploading an image! This makes it a lot clearer for the user that
the app is doing something in the background during that time!

2) Removed a weird bug when switching from Gallery to Profile, where
when the last selected sphere (ie. the current background image)
swapped its texture, it would then take on the new texture (without the
user selecting the ImageSphere. I spotted that the code paths were
different and that in the Profile image setting, I was passing the
byte[] directly to the ImageSphere. So I changed it to use the same
code path as the gallery, where I’m passing a Texture2D (not a byte[])
and it seems to have removed the bug. A new bug may arise at some point
when I get help with removing the synchronous texture construction
block, but I’ll deal with that then!