Added a Transparent Background to all button icons in the app, so they are always visible regardless of the background's colour!
I added an image component with the TransparentBackground texture to
the VRImageButton prefab. It doesn’t need to be active all the time,
but I’ve made it active all buttons in the scene. This means that I can
see the icons regardless of the colours of the VR image on my
background sphere!