Added some custom analytics + NewUserText on the Profile screen
In terms of Custom Events to Mobile Analytics I’m not yet sure exactly
what will be best to keep track of, for now I’ve simply added analytics
to track all the main button presses. So there’s an event for when
every button gets pressed: Profile, Gallery, ImageSphere, Next,
Previous, Upload.

I also added some NewUserText which explains to the user that shows up
when I validate that the user has yet to upload any 360-images to their
profile. The text will suggest to the user that they should upload some
images to get started, and it will also explain to them that swiping
up/down can hide/show the menu!

In doing this I also spotted a small bug that I fixed, which was that
the Profile list of image paths was not getting cleared, so every time
you clicked on Profile you would be adding the file paths to a list
already full of paths.