Set App's primary background + Loading Sphere images + Cleaned up some code
Sphere’s will now come up with a LoadingSphere when in the process of
Loading the first Images in the Gallery or Profile. This makes it clear
that there’s images on the way, because before it wouldn’t always be
obvious that anything was happening with the app.

My primary background for the app is now one I’ve got from Nick
Hobgood’s profile on Flickr and its credited to him with some text I’ve
written at the bottom.

I’ve also done a bit of a tidy of some code throughout and stuck a few
TODO’s for Error handling, which is the main thing left to do for V1
before releasing, other than the dreaded Synchronous block bug!