Added the option for Users to LogOut - need to make the button more hidden though
This took more work than you would expect because I found out that…

1) My Coroutine.Clear() function wasn’t actually clearing the currently
running Coroutine, only the remaining ones in the queue. This was
causing an issue with sphere’s appearing in the login screen when you
went from Profile to LogIn state. But its now fixed.

2) Facebook’s Unity SDK actually caches of all your login details
between sessions of the app (even if you close the app entirely), so
you don’t need to Login every time, as I previously had it.

The issue above also has another very significant impact which is that
you actually need to LogOut on the VReel app (not on the Facebook app,
in fact you can even delete the Facebook app and it will still allow
you to login to VReel because of the cached Facebook login values)
before you can switch between users! So I had to add a LogOut button in
order to allow switching between users on a single device!

Now you can correctly switch between multiple users on a single device,
through Logging out in VReel and Facebook, and logging in with a new
user on Facebook then VReel. And this works correctly with Downloading
different images from the S3 bucket, which I tested by switching
between mine and my Mum’s Facebook account.

The only thing I don’t like about my current setup is that the LogOut
button is really obvious to the user and very appealing to press. I
need to put it in a less obvious and less appealing part of the screen!