Reduced (but not removed) the synchronous blocking on the main thread...
I’ve been working on the Synchronous Block issue with Unity’s Tech
Support in a seperate public project on GitHub so that they have access
to it.

Together we created a C++ plugin (that I’ve exported to
libandroidcppnative.so) and a wrapper class in C# (called
CppPlugin.cs), which performs the image decoding on a background thread.

The plugin and C++ code itself makes use of a “stb_image.h”, a
fantastic and entirely open/free to use library for loading images, and
essentially just calls through to this library’s functions.

In order to get threading to work, I also created a new class called
ThreadJob.cs, which encapsulates the running of a function on a
background thread by calling Start() with the function as a parameter,
and you can quite happily sit in a co-routine loop until its finished
running the function by calling WaitFor().

AWSS3Client.cs and DeviceGallery.cs now have their own instance of a
ThreadJob and CppPlugin each, and simply call through to CppPlugin
appropriately (in AWSS3Client.cs with a pointer to memory, and in
DeviceGallery.cs with a file path), passing through their own
background ThreadJob for the CppPlugin code to use when decoding the
Image!

—————

In order to build and update the C++ plugin there are a few steps to do:
(1) Open the CppPlugin folder as a project in Android Studio 2.2
(2) Click on the Build button (or press F9), after making code changes
(3) Find the newly generated “libandroidcppnative.so” in
“CppPlugin/app/build/intermediates/cmake/debug/obj”, and copy them into
their appropriate folders into Unity in “Assets/Plugins/Android/libs”.

—————

The C++ Plugin has certainly made a substantial improvement in reducing
the frame outs and the overall problem, since its moved a lot of the
offending slow code over to a background thread.

However, we still haven’t been able to move the construction of the
Texture2D, or the copying from one Texture2D to another onto a
background thread, and these ultimately still lead to noticeable and
unpleasant frame outs =(

Moreover, the constant throwing away of old Texture2D’s and subsequent
Garbage Collection (with Resources.UnloadUnusedAssets() being called),
is also still leading to a frame out, which is particularly noticeable
when switching between Profile and Gallery where all 5 spheres have
their images swapped for the loading image.

Perhaps the way to go is to create a Texture2D at the start which is
large enough to handle any image I throw at it, then just keep setting
its pixels individually. And if this causes an issue because we can see
the Texture2D being written to live, then I can sort of double buffer
each ImageSphere, and flip between images when the writing has finished…