Repositioned Sign Out button + Re-added original Unity image decoding code
I moved the Sign Out button down to try and hide it a little bit, its
now below all the spheres. I also moved the text crediting Nick Hobgood
down on the jpg so that it didn’t get in the way of the button.

Unfortunately, the Plugin code has caused some undesirable side-effects
to functionality, so I actually reverted to using the old pre-plugin
code even though its got a worse frame-rate.

The plugin was unfortunately flipping all images on their y-axis
(something that was not obvious in the Test case where the image is
pretty symmetrical). And the C# code was running into an Out of Memory
exception as soon as you pressed the Right Arrow and moved along to the
next set of images. I’ll need to fix these issues before re-instating
it.