Removed the y-axis flip in the plugin + some incorrectly added files from the CppPlugin project
In my TransferPixelsFromSrcToDest() function I split the whole process
of scanning through the pixels into width-long scan lines, then went
through each scan line in reverse order.

With regards to reading memory this means that we now read from the
very bottom upwards (as we were doing before), but then we minus the
width of a scan line and read forwards towards the end of the scan line!

This removed the problem and meant I didn’t need to do any hacks, like
flipping the ImageSphere itself on the y-axis (which was the approach I
originally took to solve the problem!).