We now handle Loss of Internet Connection
I had a look at how some other apps handle Losing the Internet: Pokemon
Go simply blocks all UI whilst there’s no Internet Connection, whilst
Instagram allows you to press on the different tabs, but essentially
none of them do anything.

I decided that the easiest thing to do was to follow Pokemon Go’s
method and simply block all functionality and the UI, whilst you had no
Internet connection, because although you can still view all your
images in your Gallery its essentially pointless to do so without a
connection.

So I downloaded a script from the Asset Store (for $18) called
InternetReachabilityVerifier. And with this script makes it was very
easy for me to check whether we are connected to the internet or not.
So  I now have a function in the AppDirector called
VerifyInternetConnection() that can be ran from a co-routine to check
and block the UI up until the internet connection returns.

I’ve stuck a call to VerifyInternetConnectionI() at the top of any
function that requires speaking to the outside world - these are mainly
in UserLogin.cs and AWSS3Client.cs!