Fixed Analytics of SphereNumber
I had SphereNumber being passed in as a Metric however that makes more
sense for things like TextureWidth and TextureHeight. In reality the
Sphere Number is a discrete piece of data, that I need to have to
differentiate between which spheres are being pressed.

This was a simple fix, but a good thing to do, as its also clarified
the difference between Attributes and Metrics in terms of Analytics