Deleted the folders "VReal_Library_Android" and "VReal_Library_Mac"
These folders are entirely legacy and I had them when I previously was
migrating between Mac and Android BuildSettings early on in the project
and it took forever to do the operation.

However, now I never run on Mac, only on my Android phone, and the
folders were unnecessarily large and pointless!