Replaced the C++ Plugin with the one that I've been building in my Texture2DTest
After slightly adapting the code in this plugin, it should in principle
allow me to do Asynchronous Texture Loading for VReel, as it has been
doing in my Texture2DTest!

I’m submitting the code without actually replacing the “.so” in my
project or getting it to work properly, so that I keep track of the
original version of the plugin, that was indeed working in my
Texture2DTest, before I make the necessary modifications for it to work
in VReel.

For the sake of chronicling my changes, the C++ Plugin has been built
to work with the following function call order:

(1) Init() - calls glGenTextures() and allocates memory;
(2) LoadIntoWorkingMemoryFromImagePath() - calls through to stbi_load()
and sets pixels into m_pWorkingMemory
3) CreateEmptyTexture() = calls glTexImage2D() hence allocating the
actual texture
(4) LoadScanlinesIntoTextureFromWorkingMemory() - called repeatedly
until all scanlines are uploaded to the texture through glTexSubImage2D
(5) CreateExternalTexture() - called with the texture that’s been
created beneath us!
(6) Terminate() - glDeleteTextures() and delete[] alloc’d memory

Functions (1), (3) and (4) run OpenGL code and hence need to be ran on
Unity’s Render Thread (so that they are in the correct context), and so
they are called in C# as such:
”GL.IssuePluginEvent(GetRenderEventFunc(),
(int)RenderFunctions.kInit)”, where RenderFunctions is an Enum
containing the function names {kInit, kCreateEmptyTexture,
kLoadScanlinesIntoTextureFromWorkingMemory, kTerminate}.

Function (2) doesn’t require OpenGL’s participation, so it simply runs
outside on the ThreadJob class that makes use of C#’s ThreadPool, in
order to do the work on a background thread.

Function (5) is entirely in the C# side of things, and constructs a
Texture2D by using a pointer to the texture we created through OpenGL
in the C++ Plugin!