Removed the Texture Sync bug!!
The previous issues of the Skybox sphere needing to be able to show a
different image to the other spheres was fixed not by copying textures
but instead by creating a Texture Pool (something that Owen Angell
suggested, especially as he said it would be a bad idea to copy large
textures).

We have an int array called “m_textureIndexUsage” within the
m_imageSphereController (who is now solely responsible for access to
the C++ plugin and the underlying OpenGL textures). And using this int
array we essentially reference count the underlying textures in C++’s
plugin, such that spheres (including the main skybox sphere) are
allowed to share the same texture but not allowed to take a texture
away from another. Since image loading is serialised (ie. not done in
parallel), we only actually need 1 spare texture, but I’ve given an
extra one, so we only have 6 + 2 = 8 textures.

I also discovered that there’s no problem with the LoadingTexture (as I
previously believed) and hence no more issues in terms of a Synchronous
Texture block!! But there is a block caused by the OpenAndroidGallery()
function, which I need to take a look at next!