Moved UI around such that all spheres are in view when looking straight ahead
I’ve watched people use the app countless times and miss the fact that
the spheres are updating with new 360 images, because they are looking
dead ahead but the spheres are off in a slightly peripheral view.

Although I really liked having the UI around you so that users are
forced to look around a bit more, I think its probably best for now to
just have it all in view from the start.

I also updated the Upload button, to one that looks more globally
recognised as an Upload button.