Set the Project to have "Development Build" off + Wrapped all Debug.Log() code around "if (Debug.isDebugBuild)"
No debug logging should now be called within the release version of the
app.

Doing this required changing a few pieces of code here and there, in
particular this is because any function that runs on a separate thread
cannot check Debug.isDebugBuild for example.

I also removed an old script called VRCameraController that only worked
in the Editor, but the project no longer works in the editor (and has
not done for a while since I started having plugins) so its unnecessary!