Re-instated the LoginFlow, and created a BackEndAPI class to wrap comms with the Back-end + I also updated the KeyStore and submitted it
I managed to build the basic login flow with the login screens working,
the VR keyboard coming up, and the code all but communicating correctly
with the vreel-staging API in a new file called LoginFlow (I’d deleted
this concept when I’d gone with FB login).

I also built the basic framework for how I will communicate with the
API in a class I called BackEndAPI.cs. This was all done in a separate
scene so the original scene is unharmed and unchanged. But the idea
will be to delete a bunch of the old stuff as soon as I get the code
fully switched to using the BackEndAPI class.

I also ran into some issues with signing the app and so re-created the
keystone and certificate/key and have submitted it to GitHub as I never
want to lose it, because once I’ve submitted the app to the Android
Store once, I will need the same keystore to sign and update the app
from then on!