Created the full Login/Signup/Logout flow!
Created a Profile class + User class + Added PasswordText.

The Profile class will manage the entire profile page for the user and
replace all usage of the AWSS3Client, although its still pretty bare,
with most of its innards being a copy of AWSS3Client. Although I am
using it to allow the user to logout.

The User class simply maintains all the relevant variables for a user
who’s logged in at any given time. This is a good way of centralising
the user variables, without needing for it to live in BackEndAPI.cs or
Profile.cs.

Finally, I also spotted a bug with the way that my Keyboard was doing
passwords, as it was simply splatting asterixes over the text that had
been typed in. So I created a PasswordText class that stores the actual
string, and writes the asterixes into the Text field, meaning I’m not
losing the text the user typed in.

Therefore, now we’ve got the full ability to Login as a user, Sign up
as a new user, and Logout of the current user.

The next thing that I need to get working is the Profile page showing
the thumbnails of the User who’s logged in, and then the full images!