Got Uploading all but working, except for creating the Thumbnails!
The UploadImage() function correctly lives in DeviceGallery now, and
does all the correct steps for Uploading except the creation of
Thumnbails.

The annoying thing is that I can’t get the Thumbnail creation to work
because I can’t call the functions in the System.Drawing namespace, due
to it looking for a gdiplus.dll and not finding it - I added a dll for
this but the runtime doesn’t seem to find it. =(

I’m still checking this in, because once I’ve got the Thumbnails
getting created, then this will be functional.