Got the Downloading of the Original image when ImageSphere is clicked working!
ImageSphere’s simply had the Thumbnails of the post on them.

Now when the user selects a sphere I’m actually:
(1) Refreshing the data associated with the Post by calling the API and
updating both the ThumbnailURL and OriginalURL (Because they timeout
and need updating)
(2) Downloading the image associated with the OriginalURL and uploading
it through my usual code path that eventually sets it on the Skybox

In this process I changed a field that the ImageSphere holds, which is
that  it no longer always holds a filePath (this was useless for urls).
Now it holds an Identifier, which is either (1) a local device path, or
(2) a post Identifier. The post identifier can then be used later to
grab related post data (as is done when grabbing the OriginalURL).