Got a temporary Java-based solution for creating a Thumbnail
The method is only temporary because its synchronous and slow, but I
think that I might be able to write some Java code that’s much more
efficient later on from what I’ve seen on Stack Overflow.

In the process I also tidied the CppPlugin and turned it into a
component because my current implementation of the low level C++ Plugin
requires the functions to be done one at a time, so there’s no point
having multiple of them, seems safer to just have a single one. Plus as
a component I could move the concept of Texture Indexes into the
component taking it away from the ImageSphereController (which I didn’t
think should have any knowledge of texture indices…)