Tidied many bits and bobs related to CppPlugin
Since the CppPlugin is now its own component (which is wrong in fact,
as there should probably be an ImageLoader component, that internally
makes use of CppPlugin), and that I added a CoroutineQueue (so that it
wouldn’t ever trample over itself), I needed to worry about new issues.

1) I had to move the downloading of Image Data from a URL into
CppPlugin because otherwise I would pass streams that had already been
closed.

2) I needed to make sure it stopped loading at the relevant times - for
example when Logging out. Before this was done by telling the
Profile/Gallery to stop Loading, but now I need to also ask the
CppPlugin to do so otherwise it would still load the Spheres on the
Profile screen!

3) I also ran into issues with my TextureIndexing because we could get
into the case where ImageSphere’s would register a texture as being in
use from a SetImage call but never unregister it if issued another
SetImage call - this is apparent because in fact we now have really
fast loading of Images going on in the Profile (before the slow loading
meant we never ran into this problem). I fixed this by tracking the
Texture an ImageSphere is supposed to transition to properly!

There’s still so much work to do now though. In particular there’s the
following:
1) Handle when a user has too many images for a single Posts_GetAll()
call.
2) Convert CppPlugin into ImageLoader, and re-assign responsibilities
(ImageSphereController is incorrectly acting as a pathway).
3) Cache full images that are downloaded
4) Remove all frame-outs that have been newly introduced.
5) Speed up Gallery (see if I can do some Thumbnail magic)
6) Add Forgot-Password and Delete image calls.
7) Fix any quirky bugs, like Keyboard buttons getting stuck on a
pressed state.